error fixed (client)

+++ DKU_Server/DKU_ServerCore/CommonDefine.cs
@@ -12,6 +12,6 @@ namespace DKU_ServerCore
         public const int MAX_CONNECTION = 50;
         public const int SOCKET_BUFFER_SIZE = 4096;
         public const int COMPLETE_MESSAGE_SIZE_CLIENT = 1024 * 20;
-        public const Int16 HEADER_SIZE = 4;
+        public const int HEADER_SIZE = 4;
     }
 }
